Refactor Connect Four extension: Separate game logic, AI logic, and webview content into distinct modules; enhance game state management; improve AI decision-making with Minimax algorithm; update webview for better user interaction.
